refactor(spec): split layer structural checks by form family
Extract rule-family, ribbon orientation, and computed-y-mapped checks from
validate-structure-layers into dedicated modules with shared channel-error
helpers. Keep REQUIRED_CHANNELS next to its consumer loop. Public surface
stays layerStructuralErrors via the structure barrel.
