+++ packages/spec/src/validate-structure.ts
@@ -2,7 +2,7 @@
  * Data-free structural grammar checks (tier-2, opt-in via validate options).
  *
  * Implementation:
- *  - validate-structure-layers.ts — geom channels, rule forms, M2 layer rules
+ *  - validate-structure-layers.ts — geom channels, rule/ribbon/computed-y forms, M2 layer rules
  *  - validate-structure-scales.ts — named color/fill scheme vs family
  *  - validate-structure-style-bins.ts — binned style breaks/domain well-formedness
  *  - validate-structure-facet.ts — wrap XOR grid form
